+++ src/download.js
@@ -1,29 +1,9 @@
 const fs = require('fs');
 const stream = require('stream');
 const { pipeline } = require('stream/promises');
-const ProgressBar = require('progress');
 
 const { fatal } = require('./utils/logging');
-
-const MB_BYTES = 1024 * 1024;
-
-const progressStream = function (total, tokens) {
-  var pt = new stream.PassThrough();
-
-  pt.on('pipe', function (stream) {
-    const bar = new ProgressBar(tokens, { total: Math.round(total) });
-
-    pt.on('data', function (chunk) {
-      const elapsed = new Date() - bar.start;
-      const rate = bar.curr / (elapsed / 1000);
-      bar.tick(chunk.length, {
-        mbRate: (rate / MB_BYTES).toFixed(2),
-      });
-    });
-  });
-
-  return pt;
-};
+const { progressStream } = require('./utils/download');
 
 const write = fs.createWriteStream(process.argv[3]);
